requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
requests==2.32.3
Pillow==11.2.1
openai==1.93.0
huggingface-hub==0.32.2
